requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
decorator==4.4.2
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-gunicorn==0.1.1
django-heroku==0.3.1
django-phonenumber-field==7.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
facebook-sdk==3.1.0
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.4
imageio==2.22.0
imageio-ffmpeg==0.4.7
inflection==0.5.1
itypes==1.2.0
MarkupSafe==2.1.1
packaging==21.3
phonenumbers==8.12.55
phonenumberslite==8.12.55
Pillow==9.2.0
proglog==0.1.10
protobuf==4.21.6
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.1
twilio==7.14.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
